created last tests

+++ week-09/day-01/exercise-01/src/main/java/com/greenfox/cargo/Ship.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class Ship {
-  private int caliber25;
-  private int caliber30;
-  private int caliber50;
+  private double caliber25;
+  private double caliber30;
+  private double caliber50;
   private double shipStatus;
   private boolean ready;
 
@@ -18,27 +18,27 @@ public class Ship {
     this.ready = false;
   }
 
-  public int getCaliber25() {
+  public double getCaliber25() {
     return caliber25;
   }
 
-  public void setCaliber25(int caliber25) {
+  public void setCaliber25(double caliber25) {
     this.caliber25 = caliber25;
   }
 
-  public int getCaliber30() {
+  public double getCaliber30() {
     return caliber30;
   }
 
-  public void setCaliber30(int caliber30) {
+  public void setCaliber30(double caliber30) {
     this.caliber30 = caliber30;
   }
 
-  public int getCaliber50() {
+  public double getCaliber50() {
     return caliber50;
   }
 
-  public void setCaliber50(int caliber50) {
+  public void setCaliber50(double caliber50) {
     this.caliber50 = caliber50;
   }
